+++ multiversx_sdk_cli/localnet/genesis_smart_contracts_json.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from multiversx_sdk_cli.localnet import genesis
-from multiversx_sdk_cli.localnet.config import ConfigRoot
+from multiversx_sdk_cli.localnet.config_root import ConfigRoot
 
 
 def patch(data: Any, config: ConfigRoot):
