+++ multiversx_sdk_cli/localnet/step_clean.py
@@ -2,7 +2,7 @@ import logging
 from typing import Any
 
 from multiversx_sdk_cli import utils
-from multiversx_sdk_cli.localnet.config import ConfigRoot
+from multiversx_sdk_cli.localnet.config_root import ConfigRoot
 
 logger = logging.getLogger("localnet")
