+++ multiversx_sdk_cli/localnet/step_prerequisites.py
@@ -5,7 +5,7 @@ from typing import Any
 
 from multiversx_sdk_cli import dependencies, downloader
 from multiversx_sdk_cli.errors import KnownError
-from multiversx_sdk_cli.localnet.config import ConfigRoot
+from multiversx_sdk_cli.localnet.config_root import ConfigRoot
 from multiversx_sdk_cli.localnet.config_software import SoftwareResolution
 
 logger = logging.getLogger("localnet")
